+++ client/app/src/main/java/com/microsoft/research/karya/ui/payment/transactions/PaymentTransactionViewModel.kt
@@ -1,5 +1,6 @@
 package com.microsoft.research.karya.ui.payment.transactions
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.microsoft.research.karya.data.manager.AuthManager
@@ -47,7 +48,7 @@ constructor(private val authManager: AuthManager, private val paymentRepository:
               val userDate = userDateFormat.format(transactionDate)
 
               UserTransactionDetail(
-                amount = transaction.amount.toFloat(),
+                amount = transaction.amount,
                 utr = transaction.utr,
                 date = userDate ?: "",
                 status = transaction.status
